Move the options to own file
